added parseServerHost; new file for parseServer;
