+++ .server.conf
@@ -49,6 +49,12 @@ mime_add ["js"] "text/javascript"
 #
 # Can have multiple server blocks
 server {
+# Sets the ipv4 address of the server.
+#
+# One per server block, required.
+# If set to localhost, it's assigned as 127.0.0.0.
+host 42.42.42.42
+
 # Sets the port to listen to for this server.
 #
 # One per server block, required.
