+++ src/ParserUtils.cpp
@@ -1,0 +1,25 @@
+#include "ParserUtils.hpp"
+
+int	ParserUtils::atoi(std::string str) {
+	int	value = 0;
+	for (int i = 0; str[i] != '\0'; i++) {
+		if (str[i] < '0' || str[i] > '9') return (-42);
+		else { value *= 10; value += (str[i] - 48); }
+	}
+	return (value);
+}
+
+std::string	ParserUtils::removeDelimiters(std::string str) {
+	str = str.substr(1);
+	str = str.substr(0, str.length() -1);
+
+	return (str);
+}
+
+int	ParserUtils::countCharOccurs(char needle, std::string stack) {
+	int	count = 0;
+	for (int i = 0; stack[i] != '\0'; i++)
+		if (stack[i] == needle)
+			count++;
+	return (count);
+}
